requirements.txt
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ghp-import==2.1.0
griffe==0.45.2
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-autorefs==1.0.1
mkdocs-get-deps==0.2.0
mkdocstrings==0.25.1
mkdocstrings-python==1.10.3
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pymdown-extensions==10.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
urllib3==2.2.1
watchdog==4.0.1
zeep==4.2.1
